pyproject.toml
[project]
name = "audiobooks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.135.2",
    "ffmpeg-python>=0.2.0",
    "ipykernel>=7.2.0",
    "pymupdf>=1.27.2.2",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.42.0",
]
